+++ python/tvm/meta_schedule/__init__.py
@@ -35,4 +35,3 @@ from .search_strategy import (
 )
 from .tune import tune_te, tune_tir, tune_relay
 from .tune_context import TuneContext
-from .search_strategy import MeasureCandidate
