+++ python/tvm/meta_schedule/testing/__init__.py
@@ -16,5 +16,4 @@
 # under the License.
 """Testing utilities in meta schedule"""
 from .local_rpc import LocalRPC
-from .relay_workload import get_network
 from .relay_workload import MODEL_TYPE, MODEL_TYPES, get_network, get_torch_model
